+++ Changelog.md
@@ -3,7 +3,8 @@
 Features:
  * Inline Assembly: Show useful error message if trying to access calldata variables.
  * Inline Assembly: Support variable declaration without initial value (defaults to 0).
- * Type Checker: Disallow value transfers to contracts without a payable fallback function
+ * Type Checker: Disallow value transfers to contracts without a payable fallback function.
+ * Type Checker: Raise proper error for arrays too large for ABI encoding.
 
 Bugfixes:
  * Type Checker: Fix invalid "specify storage keyword" warning for reference members of structs.