+++ Changelog.md
@@ -1,6 +1,7 @@
 ### 0.4.14 (unreleased)
 
 Features:
+ * C API (``jsonCompiler``): Export the ``license`` method.
  * Inline Assembly: Show useful error message if trying to access calldata variables.
  * Inline Assembly: Support variable declaration without initial value (defaults to 0).
  * Type Checker: Disallow value transfers to contracts without a payable fallback function.